Split librfu into separate C files in its own directory, rfu_initializeAPI matching except for register differences
