+++ ld_script.txt
@@ -244,8 +244,10 @@ SECTIONS {
         src/agb_flash_1m.o(.text);
         src/agb_flash_mx.o(.text);
         src/siirtc.o(.text);
-        src/librfu.o(.text);
+        src/librfu_stwi.o(.text);
+        src/librfu_intr.o(.text);
         asm/librfu_intr.o(.text);
+        src/librfu_rfu.o(.text);
         asm/librfu.o(.text);
         asm/libagbsyscall.o(.text);
         tools/agbcc/lib/libgcc.a:_call_via_rX.o(.text);
